+++ src/gui/gui2_canvas.cpp
@@ -11,13 +11,6 @@ GuiCanvas::~GuiCanvas()
 {
 }
 
-void GuiCanvas::focusElement(GuiElement* element){
-    if (focus_element)
-        focus_element->focus = false;
-    focus_element = element;
-    if (focus_element)
-        focus_element->focus = true;
-}
 void GuiCanvas::render(sf::RenderTarget& window)
 {
     sf::Vector2f window_size = window.getView().getSize();
@@ -37,7 +30,7 @@ void GuiCanvas::render(sf::RenderTarget& window)
         click_element = getClickElement(mouse_position);
         if (!click_element)
             onClick(mouse_position);
-        focusElement(click_element);
+        focus(click_element);
     }
     if (InputHandler::mouseIsDown(sf::Mouse::Left) || InputHandler::mouseIsDown(sf::Mouse::Right) || InputHandler::mouseIsDown(sf::Mouse::Middle))
     {
@@ -107,6 +100,14 @@ void GuiCanvas::onKey(sf::Event::KeyEvent key, int unicode)
 bool GuiCanvas::onJoystickAxis(const AxisAction& axisAction)
 {
     return false;
+}
+void GuiCanvas::focus(GuiElement* element)
+{
+    if (focus_element)
+        focus_element->focus = false;
+    focus_element = element;
+    if (focus_element)
+        focus_element->focus = true;
 }
 
 void GuiCanvas::unfocusElementTree(GuiElement* element)