+++ src/gui/gui2_canvas.h
@@ -25,9 +25,10 @@ public:
     virtual void onKey(sf::Event::KeyEvent key, int unicode);
     virtual bool onJoystickAxis(const AxisAction& axisAction);
 
+
+    void focus(GuiElement* element);
     //Called when an element is destroyed in this tree. Recursive tests if the given element or any of it's children currently has focus, and unsets that focus.
     void unfocusElementTree(GuiElement* element);
-    void focusElement(GuiElement* element);
 };
 
 #endif//GUI2_CANVAS_H