+++ src/screenComponents/warpControls.cpp
@@ -53,8 +53,10 @@ void GuiWarpControls::setTargetSpaceship(P<PlayerSpaceship> targetSpaceship){
 void GuiWarpControls::onDraw(sf::RenderTarget& window)
 {
     // Update the label with the current warp factor.
-    if (target_spaceship)
+    if (target_spaceship) {
         label->setValue(string(target_spaceship->current_warp, 1));
+        slider->setValue(target_spaceship->warp_request);
+    }
 }
 
 void GuiWarpControls::onHotkey(const HotkeyResult& key)